Bug fix: mask memory explosion, add vm option,
previously suffered from huge memory allocation during xr.where(mask>0) for new param arrays. This was because I'd forgotten to rename the dims in the new dataarrays and it spawned new dims. Now recitified by  correctly naming dims conisstently with mask dims on creation. Also added option for local or vm in template. This changes the date extraction in the interpolation function to account for different paths.